net/colo.sh: Add cleanup for SECONDARY_DISK_MOUNT_POINT

+++ net/common.sh
@@ -112,7 +112,8 @@ clear_config_dir() {
 SECONDARY_DISK_MOUNT_POINT=/mnt/extra-disk
 setup_secondary_mount() {
   # If there is a secondary disk, symlink the config/ dir there
-  if [[ -d $SECONDARY_DISK_MOUNT_POINT ]]; then
+  if [[ -d $SECONDARY_DISK_MOUNT_POINT ]] && \
+    [[ -w $SECONDARY_DISK_MOUNT_POINT ]]; then
     mkdir -p $SECONDARY_DISK_MOUNT_POINT/config
     rm -rf "$SOLANA_CONFIG_DIR"
     ln -sfT $SECONDARY_DISK_MOUNT_POINT/config "$SOLANA_CONFIG_DIR"
